Updated Code further to Debug
push function is now overloaded to take both int and strings. If it is a string it will take the item and push it directly onto state 0-2. These states need to be changed to their proper states since the ones currently from 0-2 are arbitrary and logically do not make sense for the moment.